upload a title and desc via cli
