+++ index.js
@@ -1,105 +1,43 @@
 #!/usr/bin/env node
 
-const args = require("yargs").argv;
-const ipc = require("node-ipc").default;
+const lib = require('./lib')
+const { program } = require('commander');
 
-ipc.config.id = "replayable-cli";
-ipc.config.retry = 1500;
-ipc.config.silent = true;
-ipc.config.maxRetries = 0;
-// args:
-// private=anything (boolean) - don't expose share link
-// format=(link|md) - return format
-// image=(gif|png) - embedded image format, only if md is true
+program.name('replayable')
+  .description('Upgrade your bug reports, pulls, and readmes with clips from local development')
+  .version('0.0.9');
 
-const pngUrl = function (data) {
-  let url = `${data.API_ENDPOINT}/replay/${
-    data.replay.id
-    }/screenshot${shareQuery(data)}`;
-  return url;
-};
+program.showHelpAfterError();
 
-const gifUrl = function (data) {
-  let url = `${data.API_ENDPOINT}/replay/${data.replay.id}/gif${shareQuery(
-    data
-  )}`;
-  return url;
-};
+program.command('create', { isDefault: true })
+  .description('Create a replay and output the resulting url or markdown. Will launch desktop app for local editing before publishing.')
+  .option('-t, --title <string>', 'Title of the replay. Automatically generated if not supplied.')
+  .option('-d, --desc [body]', 'Replay description body. This may be piped in.')
+  .option('-p, --private', 'Do not expose secret share key in response.')
+  .option('--link', 'Return a link instead of markdown.')
+  .option('--png', 'Return a static png instead of gif. Only applies if -l is not supplied.')
+  .action(async function (str, options) {
 
-const imageUrl = function (data) {
-  if (args.image == "png") {
-    return pngUrl(data);
-  } else {
-    return gifUrl(data);
-  }
-};
-
-const shareLink = function (data) {
-  return `${data.WEB_URL}/replay/${data.replay.id}/${shareQuery(data)}`;
-};
-
-const shareQuery = function (data) {
-  return args.private ? "" : `?share=${data.replay.shareKey}`;
-};
-
-const markdownPreview = function (data) {
-  return `[![${data.replay.title}](${imageUrl(data)})](${shareLink(data)})
-  
-Watch [${data.replay.title}](${shareLink(data)}) on Replayable`;
-};
-
-const connectToIpc = function () {
-  return new Promise((resolve, reject) => {
-
-    ipc.connectTo("replayable");
-    ipc.of.replayable.on("connect", resolve)
-    ipc.of.replayable.on("error", (e) => {
-      if (e.code === 'ENOENT') {
-        console.log('Could not connect to Replayable Desktop App. Is it running?')
-        console.log('You may need to download and install the app from https://bit.ly/3erITXJ')
-      }
-    })
-    ipc.of.replayable.on("disconnect", function () {
-      console.log("disconnected from replayable".notice);
-    });
-  });
-}
-
-const createReplay = function () {
-  return new Promise((resolve, reject) => {
-
-    ipc.of.replayable.on(
-      "upload", //any event or message type your server listens for
-      function (data) {
-        // this is essentially the return value, as this echos to cli
-        if (args.format == "link") {
-          resolve(shareLink(data));
-        } else {
-          resolve(markdownPreview(data));
-        }
-      }
-    );
-
-    ipc.of.replayable.emit("create", {
-      title: 'abc',
-      description: 'xyz\n123'
-    });
-
-  });
-
-};
-
-if (!module.parent) {
-  (async () => {
-    await connectToIpc()
     try {
-      let result = await createReplay();
+      let result = await lib.createReplay({
+        title: this.opts().title,
+        description: this.opts().desc,
+        private: this.opts().private,
+        link: this.opts().link,
+        png: this.opts().png
+      });
       console.log(result);
     } catch (e) {
       console.log('Error: ', e)
     }
     process.exit(0);
-  })();
-}
 
-module.exports = { createReplay };
+  });
+
+
+program.parse();
+
+
+if (module.parent) {
+  module.exports = lib
+}
